+++ svd-rs/src/access.rs
@@ -46,7 +46,7 @@ impl Default for Access {
 }
 
 impl Access {
-    /// Parse a string into an [Access] value, returning [`Option::None`] if the string is not valid.
+    /// Parse a string into an [`Access`] value, returning [`Option::None`] if the string is not valid.
     pub fn parse_str(s: &str) -> Option<Self> {
         match s {
             "read-only" => Some(Self::ReadOnly),